[wasm][memory64] Allow declared maximum outside 32-bit range
For table64 and memory64 we should allow a declared maximum outside the
32-bit range. This should only fail at runtime when actually trying to
grow that big.

See https://webassembly.github.io/memory64/core/valid/types.html#memory-types
and https://webassembly.github.io/memory64/js-api/#limits.

R=evih@chromium.org

Bug: 369904698
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6014491
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97155}

+++ src/wasm/module-instantiate.cc
@@ -2137,13 +2137,14 @@ bool InstanceBuilder::ProcessImportedTable(
   if (table.has_maximum_size) {
     std::optional<uint64_t> max_size = table_object->maximum_length_u64();
     if (!max_size) {
-      thrower_->LinkError("table import %d has no maximum length; required: %u",
-                          import_index, table.maximum_size);
+      thrower_->LinkError(
+          "table import %d has no maximum length; required: %" PRIu64,
+          import_index, table.maximum_size);
       return false;
     }
     if (*max_size > table.maximum_size) {
       thrower_->LinkError("table import %d has a larger maximum size %" PRIx64
-                          " than the module's declared maximum %u",
+                          " than the module's declared maximum %" PRIu64,
                           import_index, *max_size, table.maximum_size);
       return false;
     }
@@ -2505,11 +2506,11 @@ bool InstanceBuilder::ProcessImportedMemories(
             ImportName(import_index).c_str(), imported_maximum_pages);
         return false;
       }
-      if (static_cast<uint32_t>(imported_maximum_pages) >
+      if (static_cast<uint64_t>(imported_maximum_pages) >
           memory->maximum_pages) {
         thrower_->LinkError(
             "%s: memory import has a larger maximum size %u than the "
-            "module's declared maximum %u",
+            "module's declared maximum %" PRIu64,
             ImportName(import_index).c_str(), imported_maximum_pages,
             memory->maximum_pages);
         return false;
